Refatorando User
